moved c2a function from util.js to callToAction.js
